Fix GitModelObject adaptation to IResource
A GitModelObject returns the full absolute path to the object in the git
working tree. Using that for IWorkspaceRoot.getFile() creates a resource
with that as path, not as location. Which result in all deleted
resources having bogus paths in the Git Model of the Synchronize View.

This change maps back the git absolute path to a workspace-relative path
that can then safely be used with IWorkspaceRoot.getFile().

Adapting a GitModelObject to an IResource will work only for files
actually in the Eclipse workspace. Changing that so that it works also
for files in the working tree, but not in the workspace is _not_ fixed
by this change.

Change-Id: Ie539c3ad46f3f48ac6f01a365d20ceacd8952564
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>