Show Gitflow submenu in history view only if repo is Gitflow-enabled
Don't display the whole submenu at all if the repo has not been set up
for Gitflow.

Also enable the "Start release from..." only when the current branch is
the development branch, per what the former code did and per [1] and
[2].

Note that in the GitFlow context menu in the package explorer, the
"Start Feature" and "Start Release" menu items are enabled even if the
current branch is not the development branch. I don't know if that is
intentional and have therefore left that unchanged. Same for hotfix
branches, which according to [1] should branch off master only.

[1]
http://nvie.com/posts/a-successful-git-branching-model/#release-branches
[2] http://danielkummer.github.io/git-flow-cheatsheet/

Change-Id: Ia8ba6401a74cc1f8d6668e549957726e22e8cfaa
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>