React on IWorkingSet selections in history and git staging views
Add adapter factory definitions in plugin.xml to adapt IWorkingSet to
IHistoryPageSource and to Repository, and for ResourceMapping to adapt
to Repository. This makes the history view follow selections even on
working sets, provided all projects in the working set belong to the
same repository.

In staging view, also consider adaptations to Repository.

Bug: 485464
Change-Id: I259d6a2ba8f1e49379c110be986dc753bc108d18
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>