Test stability: GitRepositoriesView tests
Test failure in [1] claimed that no refresh job was scheduled.

The RepositoriesView schedules a refresh on its own in response to an
index diff change. If a refresh job is already scheduled and not
running yet, the re-schedule in view.refresh() has no effect. See
JobManager.schedule(). As a result, the JobListener is never notified
about a scheduled job.

Moreover, if the job is already running, re-scheduling will make it run
again, but the JobListener will return already when the currently
running job finishes.

Therefore, we just need to wait for refresh jobs after having called
view.refresh(). That will make the test wait in any case. Timeout can be
implemented via a custom progress monitor.

[1] https://hudson.eclipse.org/egit/job/egit.gerrit/8154/testReport/junit/org.eclipse.egit.ui.view.repositories/GitRepositoriesViewBranchHandlingTest/testCheckoutRemote/

Change-Id: I2ecd20b4caaf70e00a903366f46da0075afa6a3f
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>