Share project should close related editors
If the user has an editor open of a project he shares with Git, the
editor currently stays open if the repository is created in a different
path so that the sharing wizard moves the project to the repository's
location. Then the user can still use the editor but as soon as he tries
to save, he gets an error message.

This fix closes all editors of the project to be shared and saves their
content. This way no content is lost and the user is not misled to
believe he could still change these files using the old editors.

Bug: 475485
Change-Id: I0cabeb7aaa295ad652e9d298c3f513d3e9df2d2d
Signed-off-by: Lars Vogel <Lars.Vogel@vogella.com>