Make columns in Gitflow feature checkout dialog sortable
To make the dialog scale to many feature branches, it should support
sorting. With the current set of columns this makes most sense if a
a certain convention is used for either the feature names or the commit
message, such as putting the ticket ID first. In the future, this could
be used to sort by the date of the last commit on a branch.

Change-Id: I9299ef2f882a3e4ea2bd2e8eff658419a10adbd9