Harmonize Gerrit fetch and push operation availability
Determining whether a refspec is for Gerrit must be in synch with the
ResourcePropertyTester. Moreover, fetch and push refspecs must be
considered separately.

With that in place, add new properties to the property tester:
canFetchFromGerrit and canPushToGerrit. Use these properties to
determine visibility of the "Push to Gerrit..." and "Fetch from
Gerrit..." commands such that they are visible only if they can be
performed.

Bug: 493352
Change-Id: Idc66e6a912134d8979b3fad4ce6db4395388a1e2
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>