Immediate user feedback on failed push background jobs
With commit 09e0d369 we've eliminated asynchronous dialogs from
background jobs for pushing and fetching. Instead, we associate an
action to open the result dialog with the job; the user can invoke that
action from the progress view.

This works nicely, but makes it easy to overlook a failed push.

Try to improve that by making the job return an ERROR status when a
push failure has occurred. This makes the job/status framework notify
the user; the notification dialog will show the failure message and
additionally a button giving access to the job's action, which will
open the full push result dialog.

Bug: 498176
Change-Id: Ia12c8586a0bf3827fe8916d1fa5853cbfb1e7604
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>