Clean up text editors' ruler context menu
EGit mistakenly contributed the Team menu also to the context menu in
the ruler of the active text editor. Since we do want to use the
locationURI "popup:org.eclipse.ui.popup.any" for the Team menu, the only
way to ensure it isn't displayed in a ruler's context menu I see is to
explicitly exclude it from the "#AbstractTextEditorRulerContext" menu.

Also apply the same fix to gitflow's "Compare with develop" and "Replace
with develop" contributions, and also restrict those to gitflow-enabled
repositories.

Bug: 469610
Change-Id: I19c843076adb7a4c3dc83980b44866d8b2c6de0d
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>