Fix focus restoration in CommitEditor pages
FormPage by default re-sets the focus to the first child of the
managed form. In the CommitEditorPage, this is not convenient;
it'll always set focus on the author indication. Add focus tracking
to identify the control currently having the focus so that it can
be re-focused when the page becomes active again after a page
change or editor de-activation.

Because of the vagaries of focus handling in MultiPageEditorParts,
this turned out to be more convoluted than expected. Depending on
how the focus gets set (through the page or through the editor),
focus is either set via the page's SWT control directly, or
through an IFormPart registered in the ManagedForm. So provide
both mechanisms.

Bug: 509486
Change-Id: I678bdfde9105d04669a42ad81502eb1064c470bc
Also-by: Markus Keller <markus_keller@ch.ibm.com>
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>
Signed-off-by: Markus Keller <markus_keller@ch.ibm.com>