Fix NPE when trying to show diff for a merge commit
JGit cannot compute meaningful diffs for merge commits. Display a
message saying so in the DiffEditorPage instead of throwing an NPE
(which would get caught, logged, and reported to the user in an error
dialog).

The alternative would have been to not add a DiffEditorPage at all for
merge commits, but I considered that confusing. Users would be left
wondering why for some commits no diff page was present.

Bug: 460442
Change-Id: I1af0a0b1f67c765025c5c0545b53c5df375eb08c
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>