Add ICommitMessageProvider2 for caret positioning in commit messages
The caret in the CommitMessageComponent is placed according to the
value of getDesiredCaretPosition() of the first implementation of
ICommitMessageProvider2 that provides caret position. In case there
are multiple implementations available that provide a defined caret
position the second and subsequent positions are ignored.

Commit messages are not trimmed anymore to ensure consistency between
a provided message and the provided caret position.

Includes unit and UI tests.

CQ: 13790
Bug: 516867
Change-Id: I1b0ffd33ef37196d53077d09774672c2b033835f
Signed-off-by: Stefan Rademacher <stefan.rademacher@tk.de>
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>