Change the icons for clone, add, and create repositories
Those icons in the repositories view's toolbar were just too
cluttered, and the small text didn't work well. Simplify the
graphics re-using the cloud metaphor as suggested on bugzilla.

Since both the folder symbol and our cloud are yellow, add a
green arrow to the "clone" icon to distinguish it better from
the "add" icon.

The "create" icon has a yellow repository symbol and a yellow
plus, as before. A little green plus might make that plus more
noticeable, but I have decided against that because:

* in other places in Eclipse, a _yellow_ plus signifies "add
  something new", and I didn't want to deviate from that visual
  language, and
* the old "add" icon had a green plus, so users might get
  confused.

Bug: 347081
Change-Id: I6857cc62ed567fae368e86939a12ecf96d98698e
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>