Fix visibleWhen for Team->Advanced->Clean
The handler introduced in commit 23ffce5 is active when either a
RepositoryNode _or_ a Project from a git repository is selected.
The command in the Team menu thus needs to have an explicit
visibleWhen clause to show it only when projects are selected,
otherwise it also appears in the Repositories View.

Bug: 518098
Change-Id: I82e658af24438a6507c7bcb1957d0d37e5d71576
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>