Move history page actions into plugin.xml and remove default handlers
The history view's main context menu was built programmatically,
apparently because of bug 322751, which undid an earlier attempt
to do this (bug 321516).

Move the context menu back to plugin.xml. Remove the menu items from
CommitGraphTable and add a new section in the menu ("history.group")
so that we can add our built-in commands in our own group, not just
under "additions".

Basically all the history view commands had defaultHandlers. To
remove them, we need quite a lot of new dedicated handlers.

Verified that the error described in bug 322751 did not re-appear.

There should be only minor visual changes in the context menu of the
history view. We only show applicable operations, and I have slightly
re-arranged a few separators.

There is quite some duplication between handlers. With the new
IRepositoryCommit interface, it should be possible to unify many
commands, but this is postponed to future commits.

Bug: 495064
Change-Id: I73951763904f61ec1b116dc1781cbcc6b0fdb65f
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>