Fix NotHandledException for "org.eclipse.egit.ui.RepositoriesViewOpen"
The handler activation in plugin.xml looks correct: the handler is
activated for RefNode, TagNode, FileNode, and StashedCommitNode.

The implementation in RepositoriesView tries to execute the command
only for those.

The only explanation for the handler not being active I have is that
somehow the selection gets lost somewhere... So make sure we pass on
the selected element and run the command explicitly with that element
as its current selection.

Bug: 521824
Change-Id: I32f69aceceaa57e89acbd05ddeaab553f8113733
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>