Make commit viewer diff list scrollable
If a commit contains a big number of files, then the commit viewer diff
list cannot be scrolled down to see more than the very first files. This
seems to be caused by the section widget, which automatically sizes the
file list to contain every item.

By providing a size hint, the section no longer expands to contain all
items, but only consumes the available space of the editor.

Bug: 527451
Change-Id: Ie6afee43842afd0b52585945ace2924837115f2a
Signed-off-by: Michael Keppler <Michael.Keppler@gmx.de>