Improve check for uncommitted changes before cherry-pick
First, give the dialog an appropriate title. The default title
mentions rebasing, which is confusing when the user tries to
cherry-pick.

Second, change the logic for applying a cherry-pick. Don't try to
figure out up front whether there will be conflicts with uncommitted
changes. Instead, just do it, and if the operation failed because of
a dirty working tree or index, then ask the user to clean-up, and if
so redo the operation.

The behavior is then more in line with rebasing or checking out,
which also can succeed if the working tree is dirty.

Bug: 470452
Change-Id: I0ea3985ecebb8383f45f560e94f3cde1309d2570
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>