Branch name validation is slow for big repositories
Run the branch name validation in a job, which can be rescheduled with
every new character. Therefore it does no longer block on fast input,
only the very last character still has the small delay that it takes for
one round of validation.

This change is inspired by the patch at bug 352253, therefore I listed a
co-author.

Bug: 533198
Change-Id: Ib9fba9db6d3f4054b9a80662721b3e7ded11077e
Signed-off-by: Michael Keppler <Michael.Keppler@gmx.de>
Co-authored-by:  Curtis Windatt <curtis.windatt.public@gmail.com>