Fix sort order in Staging view
Change alphabetic sorting such that elements with shorter directory
names in the path sort before elements with longer directory names.

* Replace slashes by '\001' to ensure the above.
* If "file name first" is active, sort by file name _and_ path to
  ensure relative order of files with the same name in different
  directories.
* Also sort StagingFolderEntries the same way to ensure that the order
  is the same when "compact tree" representation is active.

Also apply the same technique in FileDiff.PATH_COMPARATOR to ensure
that we use a consistent ordering also in the history page and in the
commit viewer and its outline page.

Bug: 536240
Change-Id: I6a8daf7faed8a89a5b97f083e108cd614ba15096
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>