[staging] use icons for author and committer
Add the same icons as in the commit editor also to the staging area.
Reduce the unwanted margin around the group of (author, committer,
commit buttons) and reduce the gap above the commit buttons. Move
the icons and the labels closer together, and increase the spacing
between labels and the text fields by one pixel to avoid that the
content-assist lightbulb gets clipped.

Bug: 537200
Change-Id: Id70658b8776dd3b9494445be3e4913e85187f055
Signed-off-by: Michael Keppler <Michael.Keppler@gmx.de>
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>