Improve reflog view updates
Update the view only when it is visible, otherwise remember the
new input and set it when the view becomes visible.

This is the same mechanism as used in StagingView, which has been
factored out now into a new class PartVisibilityListener.

Make sure the view reacts correctly when the branch of which the
reflog is shown is deleted, or when the whole repository is deleted.

Bug: 548079
Change-Id: Id429a1e02684298a91747fa99889d4e9ebcdae9a
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>