Avoid UI blocks in ReflogView
The view already loads the reflog in the background. But the
(short) commit message was obtained through RevWalk.parseCommit
inside the label provider. This is a potentially long-running
operation, and the label provider runs in the UI thread.

Avoid this by getting the commit messages up front in the
background, too.

Change-Id: If0f33d391659aa774ac5d71f7a85dafa4792a193
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>