Indent history table columns
Let SWT paint the columns (except the one with the plot) to get native
indentation and abbreviation of texts.

Estimate the column width of the commit id column using "bbbbbbb"
instead of "0000000", since "0" is actually quite narrow in many fonts.
Move the calculation to a common utility, to also use it in the
RebaseInteractiveView.

Bug:539553
Change-Id: Ife9a56bcfd55b4dab4728a96d7d582914a1a6853
Signed-off-by: Michael Keppler <Michael.Keppler@gmx.de>