Releng: Store all project settings explicitly
We use project specific properties in many projects. However, the files
containing them are old, therefore they do not contain all the keys and
values that have been added in newer Eclipse releases.

This change makes all those missing settings explicit. All project
properties have been opened in the UI and the "Apply" button has been
used without changing any controls. We should therefore see only
additions in the diff, no changes in values.

The main intention of this change is to reduce the noise for actual
changes of settings which are intended to be applied afterwards. If that
is accepted, I will do the same on jgit and egit-github.

The removal of JDT settings in org.eclipse.egit is intentional. That is
a PDE project without Java nature.

Change-Id: I1b6d56b2f38e85ec8842076a1aa462a990d09f3f
Signed-off-by: Michael Keppler <Michael.Keppler@gmx.de>