Include the Apache MINA ssh client
Add a preference with UI in the main Git preference page to select
between JSch and the new Apache MINA sshd client.

Read preferences from the org.eclipse.jsch.core preference node and
install a listener to pick up preference changes.

The session factory thus can use the ssh preferences as configured by
the user. For the time being, we use the ssh directory, the default
identities, and the preferred authentication mechanisms. The latter
two can be overridden through the ssh config file, which will be the
one in the configured ssh directory. The other preferences from
org.eclipse.jsch are not taken into account; sshd may have different
algorithms available and I don't want to change preferences in a
foreign bundle.

Make the proxy service accessible in the EGit core activator, and use
it in a ProxyDataFactory in the EGitSshdSessionFactory.

The EGitSshdSessionFactory has no key cache. Eclipse is a long-running
application, and keeping ssh keys in memory for that long is most
probably not wise. Instead re-load keys whenever they are needed, and
use an IdentityPasswordProvider (which provides passwords for encrypted
private key files) that uses the Eclipse secure storage for key
passphrases to avoid repeatedly asking the user for the same passphrases
for the same keys.

Bug: 520927
JGit-Dependency: Iaa78bbb998a5e574fa091664b75c48a3b9cfb897
Change-Id: Id3cf850c4e132e864eab7eda52c20ff379e2b1d9
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>