Workaround for empty staging viewer after selection
Since the underlying SWT issue has been discussed since 2014 and we have
not found a real fix, have a workaround in egit instead.

On Windows add the missing redraw in the selection listener explicitly.
Users may notice a minor flickering on selection. However, that is
surely better than not seeing all the unselected elements at all. And
the flickering also exists on the first element, when NOT applying this
workaround.

Bug: 533555
Change-Id: If8b9303ca94a749fd8d80a2ec5efe07ca0a4162b
Signed-off-by: Michael Keppler <Michael.Keppler@gmx.de>