Allow to sign commits
This change adds support for signing commits to the commit dialog as
well as the staging view. Signing uses the support in JGit to set
whether a commit should be signed.

The default is inherited from the Git config value of 'commit.gpgSign'.
Toggling the state will either for JGit to sign a commit or to not sign
it.

Bug: 491169
Change-Id: Id172ff28984342970b28696b15d66cae9d92449f
Signed-off-by: Gunnar Wagenknecht <gunnar@wagenknecht.org>