Set initial focus for empty remote configuration
When the configure fetch/push dialog is invoked the very first time, it
may not be easy for users to find what to do next. Therefore focus the
"Change URI" button if the URI is empty.

This change was not directly requested in the linked bug, but after the
visual changes on the text field, this seems like a logical next step in
the sense of that bug.

Bug: 541813
Change-Id: I1fef5fab47df0d78794486f9f3b23ccfaf0b025b
Signed-off-by: Michael Keppler <Michael.Keppler@gmx.de>