Show full branch name in status bar
The status bar shows the label of the currently selected node in the
repositories view. With hierarchical layout enabled, remote branches
(and other branches with hierarchical name) can be confused with other
non-hierarchical branches, since only the last part of the hierarchical
name was shown.

With this change the full branch name is shown instead, independent of
hierarchical layout enabled or not.

Bug: 544414
Change-Id: I9159e800de5165ce0e1bca86adcd3b5256e44842
Signed-off-by: Michael Keppler <Michael.Keppler@gmx.de>