Display more icons in egit menu items
The following menu items display an icon now (they did not show an icon
before, but all the icon files already existed):
* Show Revision Information (in editor ruler)
* Show in Editor (repository view)
* Rebase Feature (gitflow)
* Checkout Feature (gitflow)
* Checkout develop (gitflow)
* Compare with > Commit
* Replace with > Commit

All other newly registered command images in this change were visible as
icons before already. They were just not bound to the command. That
means they were missing in the keys preference page and the Ctrl-3
completion, but not in the egit ui.

Change-Id: I6f82ef6c54125cd02a1f61442c5cd0a9c2e9c72f
Signed-off-by: Michael Keppler <Michael.Keppler@gmx.de>