Don't show whitespace in read-only text fields
Wherever we use the SpellcheckableMessageArea implementation in egit,
whitespace is rendered visible, if the global workspace preference is
enabled. This helps while editing a commit message, but it does not have
much value when showing a commit message of the past in the commit
viewer, or the gerrit notes in the commit viewer, or when showing the
push result message after a push operation.

Therefore initially ignore that preference for all read-only usages of
the component. The user can still enable whitespace rendering using the
context menu, and the context menu is still coupled to the global
preference (it just doesn't reflect the enabled preference initially).

Bug: 544928
Change-Id: I9dd030cbe2a0af51b4a7a058d11257f3877ed7b4
Signed-off-by: Michael Keppler <Michael.Keppler@gmx.de>