Unify "Create Branch" commands
EGit had three different commands for "Create Branch", one in the
repositories view, one in the history view, and one in the commit
editor. The one from the repositories view was the only one that
had a category ID and that thus showed up by default in the key
binding preferences.

Change that to use one single command with two different handlers,
one for RepositoryTreeNodes and one for IRepositoryCommits as used
in the history view and in the commit editor. Delete the dedicated
handler for the history view, and generalize the other such that it
works in either the history view or in the commit editor.

The command re-uses the ID from the repositories view command
because of bug bug 415382, though I would have liked to introduce
a new command ID "org.eclipse.egit.ui.team.branch.create".

Bug: 545031
Change-Id: I57f2f6c15e2f7bb796ecb3bc4c790893b094d68c
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>