Minor re-organization of FileRevisionEditorInput
Move the temp files into the plugin's state location instead of using
/tmp. Shorted file names by using only the first 7 characters of the
commit SHA1.

Also remove an unused constructor, and replace Object by IFileRevision
in the others: this is only ever used with IFileRevision.

As far as I see it is also never used with IFileState for the
storage. Removing that code is for another commit, though.

Bug: 547110
Change-Id: I88aa62a0200bd9c1b5414a93024d6649842473c5
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>