Bind F2 on repo node to rename of current branch
Add the ability to use F2 directly on the repository node to rename the
current branch instead of first having to expand all nodes to the
current branch.

It would have been possible to use the existing eclipse.ui.edit.rename
command (and just to add to the existing activeWhen expression), but
that has the bad side effect of the command still being labeled "Rename"
in Ctrl-3, when invoked on the repository. Therefore the old rename
handler has been attached to the rename branch command instead and a
separate key binding has been added to the repositry view, to override
the generic F2 rename binding there.

Bug: 545290
Change-Id: Idf7fd3fd6f9fba00954d6b7318b11182dcd4cb69
Signed-off-by: Michael Keppler <Michael.Keppler@gmx.de>